INTERLOK-3856 #comment Add java-library plugin to deal with gradle variant
The reason is that if a dependency of one of the interlok optional components is using gradle metadata with multiple variants we can get an error like:

... Cannot choose between the following variants of ...

The issue with INTERLOK-3856 is when using interlok-gcloud-pubsub. Some of its dependencies (e.g. com.google.api:gax) have a shadow (uber jar) variant of the jar file.

An alternative could have been to use configuration such as:

project.getConfigurations().configureEach( (configuration) -> {
 configuration.attributes( (attributeContainer) -> {
  attributeContainer.attribute(
    Bundling.BUNDLING_ATTRIBUTE,
    project.getObjects().named( Bundling.class, Bundling.EXTERNAL )
  });
  // plus..
  // Usage = JAVA_RUNTIME
 });
);