ci: switch to fastify github action
Remove depbot-automerge since fastify relies on dependabot-metadata
this means it has to run in a PR / PR target context rather than
workflow_call or similar.
Revert dependabot-merge back to ridedott/merge-me which is still "best"
IMO, but doesn't work because of token scope/protected branches; it
is manually marked as disabled.