cve: modify snakeyaml suppressions
We are optimistically upgraded to snakeyaml1.3.3 by jackson so remove
the 1.3.1 CVE suppressions.
Add suppression for 2022-1471 since we don't use snakeyaml directly and
have to rely on jackson to use SafeConstructor.
Modify build.gradle to include ../owasp-exclude so we are always using
at least an exclusion that is local (otherwise we can't test changes).